GA-23: renamed to UserAccessType
