+++ src/main/java/uk/gov/hmcts/reform/professionalapi/util/RefDataUtil.java
@@ -31,13 +31,14 @@ import uk.gov.hmcts.reform.professionalapi.controller.response.ProfessionalUsers
 import uk.gov.hmcts.reform.professionalapi.controller.response.ProfessionalUsersEntityResponseWithoutRoles;
 import uk.gov.hmcts.reform.professionalapi.controller.response.ProfessionalUsersResponse;
 import uk.gov.hmcts.reform.professionalapi.controller.response.ProfessionalUsersResponseWithoutRoles;
-import uk.gov.hmcts.reform.professionalapi.domain.AccessType;
 import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
 import uk.gov.hmcts.reform.professionalapi.domain.OrganisationStatus;
 import uk.gov.hmcts.reform.professionalapi.domain.PaymentAccount;
 import uk.gov.hmcts.reform.professionalapi.domain.ProfessionalUser;
 import uk.gov.hmcts.reform.professionalapi.domain.SuperUser;
+import uk.gov.hmcts.reform.professionalapi.domain.UserAccessType;
 import uk.gov.hmcts.reform.professionalapi.domain.UserAccountMap;
+import uk.gov.hmcts.reform.professionalapi.domain.UserConfiguredAccess;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -426,8 +427,8 @@ public class RefDataUtil {
                 .getUserProfiles()) {
             for (ProfessionalUser pu : professionalUsers) {
                 if (pu.getUserIdentifier().equals(professionalUsersResponse.getUserIdentifier())) {
-                    professionalUsersResponse.getAccessTypes().addAll(pu.getUserConfiguredAccesses().stream()
-                            .map(uca -> AccessType.fromUserConfiguredAccess(uca)).collect(toList()));
+                    professionalUsersResponse.getUserAccessTypes().addAll(pu.getUserConfiguredAccesses().stream()
+                            .map(uca -> fromUserConfiguredAccess(uca)).collect(toList()));
                     professionalUsersResponse.setLastUpdated(pu.getLastUpdated());
                 }
             }
@@ -439,8 +440,8 @@ public class RefDataUtil {
         for (ProfessionalUsersResponse professionalUsersResponse : professionalUsersEntityResponse.getUsers()) {
             for (ProfessionalUser pu : professionalUsers) {
                 if (pu.getUserIdentifier().equals(professionalUsersResponse.getUserIdentifier())) {
-                    professionalUsersResponse.getAccessTypes().addAll(pu.getUserConfiguredAccesses().stream()
-                            .map(uca -> AccessType.fromUserConfiguredAccess(uca)).collect(toList()));
+                    professionalUsersResponse.getUserAccessTypes().addAll(pu.getUserConfiguredAccesses().stream()
+                            .map(uca -> fromUserConfiguredAccess(uca)).collect(toList()));
                     professionalUsersResponse.setLastUpdated(pu.getLastUpdated());
                 }
             }
@@ -520,9 +521,17 @@ public class RefDataUtil {
             String invalidAddId = invalidAddIdsSet.stream().collect(Collectors.joining(", "));
             throw new ResourceNotFoundException(ERROR_MSG_ORG_IDS_DOES_NOT_MATCH + " : " + invalidAddId);
         }
+    }
 
+    public static UserAccessType fromUserConfiguredAccess(UserConfiguredAccess userConfiguredAccess) {
+        UserAccessType accessType = new UserAccessType();
+        accessType.setAccessTypeId(userConfiguredAccess.getUserConfiguredAccessId().getAccessTypeId());
+        accessType.setOrganisationProfileId(userConfiguredAccess.getUserConfiguredAccessId()
+                .getOrganisationProfileId());
+        accessType.setJurisdictionId(userConfiguredAccess.getUserConfiguredAccessId().getJurisdictionId());
+        accessType.setEnabled(userConfiguredAccess.getEnabled());
 
-
+        return accessType;
     }
